requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--index-url https://anthony.delage%40qubit.com:AP2dPP8zWYg4Ep6T1rByWEpYeoP@qubit.jfrog.io/qubit/api/pypi/qubit-python/simple

chromedriver-binary==80.0.3987.16.0
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1    # via pandas
pytz==2019.3              # via pandas
pyyaml==5.2
selenium==3.141.0
six==1.13.0               # via python-dateutil
urllib3==1.25.7           # via selenium
